+++ platformio/__init__.py
@@ -14,14 +14,16 @@
 
 import sys
 
-VERSION = (3, 5, 0)
+VERSION = (3, 5, 1)
 __version__ = ".".join([str(s) for s in VERSION])
 
 __title__ = "platformio"
-__description__ = ("An open source ecosystem for IoT development. "
-                   "Cross-platform build system and library manager. "
-                   "Continuous and IDE integration. "
-                   "Arduino, ESP8266 and ARM mbed compatible")
+__description__ = (
+    "An open source ecosystem for IoT development. "
+    "Cross-platform IDE and unified debugger. "
+    "Remote unit testing and firmware updates. "
+    "Arduino, ARM mbed, Espressif (ESP8266/ESP32), STM32, PIC32, nRF51/nRF52, "
+    "FPGA, CMSIS, SPL, AVR, Samsung ARTIK, libOpenCM3")
 __url__ = "http://platformio.org"
 
 __author__ = "Ivan Kravets"
@@ -36,5 +38,5 @@ if sys.version_info < (2, 7, 0) or sys.version_info >= (3, 0, 0):
     msg = ("PlatformIO Core v%s does not run under Python version %s.\n"
            "Minimum supported version is 2.7, please upgrade Python.\n"
            "Python 3 is not yet supported.\n")
-    sys.stderr.write(msg % (__version__, sys.version.split()[0]))
+    sys.stderr.write(msg % (__version__, sys.version))
     sys.exit(1)