+++ platformio/managers/core.py
@@ -21,9 +21,9 @@ from platformio import __version__, exception, util
 from platformio.managers.package import PackageManager
 
 CORE_PACKAGES = {
-    "contrib-piohome": ">=0.6.0,<2",
+    "contrib-piohome": ">=0.6.1,<2",
     "contrib-pysite": ">=0.1.2,<2",
-    "tool-pioplus": ">=0.12.1,<2",
+    "tool-pioplus": ">=0.13.3,<2",
     "tool-unity": "~1.20302.1",
     "tool-scons": "~3.20501.2"
 }
@@ -101,8 +101,7 @@ def pioplus_call(args, **kwargs):
         raise exception.PlatformioException(
             "PlatformIO Core Plus v%s does not run under Python version %s.\n"
             "Minimum supported version is 2.7.6, please upgrade Python.\n"
-            "Python 3 is not yet supported.\n" % (__version__,
-                                                  sys.version.split()[0]))
+            "Python 3 is not yet supported.\n" % (__version__, sys.version))
 
     pioplus_path = join(get_core_package_dir("tool-pioplus"), "pioplus")
     pythonexe_path = util.get_pythonexe_path()