+++ tests/commands/test_lib.py
@@ -15,8 +15,7 @@
 import json
 import re
 
-from platformio import exception, util
-from platformio.commands.init import cli as cmd_init
+from platformio import exception
 from platformio.commands.lib import cli as cmd_lib
 
 
@@ -36,24 +35,10 @@ def test_search(clirunner, validate_cliresult):
 def test_global_install_registry(clirunner, validate_cliresult,
                                  isolated_pio_home):
     result = clirunner.invoke(cmd_lib, [
-        "-g", "install", "58", "547@2.2.4", "DallasTemperature",
-        "http://dl.platformio.org/libraries/archives/3/5174.tar.gz",
-        "ArduinoJson@5.6.7", "ArduinoJson@~5.7.0", "168@00589a3250"
+        "-g", "install", "64", "ArduinoJson@~5.10.0", "547@2.2.4",
+        "AsyncMqttClient@<=0.8.2", "999@77d4eb3f8a"
     ])
     validate_cliresult(result)
-
-    # check lib with duplicate URL
-    result = clirunner.invoke(cmd_lib, [
-        "-g", "install",
-        "http://dl.platformio.org/libraries/archives/3/5174.tar.gz"
-    ])
-    validate_cliresult(result)
-    assert "is already installed" in result.output
-
-    # check lib with duplicate ID
-    result = clirunner.invoke(cmd_lib, ["-g", "install", "305"])
-    validate_cliresult(result)
-    assert "is already installed" in result.output
 
     # install unknown library
     result = clirunner.invoke(cmd_lib, ["-g", "install", "Unknown"])
@@ -62,9 +47,9 @@ def test_global_install_registry(clirunner, validate_cliresult,
 
     items1 = [d.basename for d in isolated_pio_home.join("lib").listdir()]
     items2 = [
-        "ArduinoJson_ID64", "ArduinoJson_ID64@5.6.7", "DallasTemperature_ID54",
-        "DHT22_ID58", "ESPAsyncTCP_ID305", "NeoPixelBus_ID547", "OneWire_ID1",
-        "EspSoftwareSerial_ID168"
+        "ArduinoJson_ID64", "ArduinoJson_ID64@5.10.1", "NeoPixelBus_ID547",
+        "AsyncMqttClient_ID346", "ESPAsyncTCP_ID305", "AsyncTCP_ID1826",
+        "RFcontrol_ID999"
     ]
     assert set(items1) == set(items2)
 
@@ -72,11 +57,12 @@ def test_global_install_registry(clirunner, validate_cliresult,
 def test_global_install_archive(clirunner, validate_cliresult,
                                 isolated_pio_home):
     result = clirunner.invoke(cmd_lib, [
-        "-g", "install", "https://github.com/adafruit/Adafruit-ST7735-Library/"
-        "archive/master.zip",
+        "-g", "install",
         "http://www.airspayce.com/mikem/arduino/RadioHead/RadioHead-1.62.zip",
         "https://github.com/bblanchon/ArduinoJson/archive/v5.8.2.zip",
-        "https://github.com/bblanchon/ArduinoJson/archive/v5.8.2.zip@5.8.2"
+        "https://github.com/bblanchon/ArduinoJson/archive/v5.8.2.zip@5.8.2",
+        "http://dl.platformio.org/libraries/archives/0/9540.tar.gz",
+        "https://github.com/adafruit/Adafruit-ST7735-Library/archive/master.zip"
     ])
     validate_cliresult(result)
 
@@ -87,16 +73,11 @@ def test_global_install_archive(clirunner, validate_cliresult,
     ])
     assert result.exit_code != 0
 
-    # check lib with duplicate URL
-    result = clirunner.invoke(cmd_lib, [
-        "-g", "install",
-        "http://www.airspayce.com/mikem/arduino/RadioHead/RadioHead-1.62.zip"
-    ])
-    validate_cliresult(result)
-    assert "is already installed" in result.output
-
     items1 = [d.basename for d in isolated_pio_home.join("lib").listdir()]
-    items2 = ["Adafruit ST7735 Library", "RadioHead-1.62"]
+    items2 = [
+        "RadioHead-1.62", "ArduinoJson", "DallasTemperature_ID54",
+        "OneWire_ID1", "Adafruit ST7735 Library"
+    ]
     assert set(items1) >= set(items2)
 
 
@@ -113,18 +94,41 @@ def test_global_install_repository(clirunner, validate_cliresult,
             "https://gitlab.com/ivankravets/rs485-nodeproto.git",
             "https://github.com/platformio/platformio-libmirror.git",
             # "https://developer.mbed.org/users/simon/code/TextLCD/",
-            "knolleary/pubsubclient"
+            "knolleary/pubsubclient#bef58148582f956dfa772687db80c44e2279a163"
         ])
     validate_cliresult(result)
     items1 = [d.basename for d in isolated_pio_home.join("lib").listdir()]
     items2 = [
         "PJON", "PJON@src-79de467ebe19de18287becff0a1fb42d",
         "ArduinoJson@src-69ebddd821f771debe7ee734d3c7fa81", "rs485-nodeproto",
-        "PubSubClient"
+        "platformio-libmirror", "PubSubClient"
     ]
     assert set(items1) >= set(items2)
 
-    # check lib with duplicate URL
+
+def test_install_duplicates(clirunner, validate_cliresult, without_internet):
+    # registry
+    result = clirunner.invoke(cmd_lib, [
+        "-g", "install",
+        "http://dl.platformio.org/libraries/archives/0/9540.tar.gz"
+    ])
+    validate_cliresult(result)
+    assert "is already installed" in result.output
+
+    # by ID
+    result = clirunner.invoke(cmd_lib, ["-g", "install", "999"])
+    validate_cliresult(result)
+    assert "is already installed" in result.output
+
+    # archive
+    result = clirunner.invoke(cmd_lib, [
+        "-g", "install",
+        "http://www.airspayce.com/mikem/arduino/RadioHead/RadioHead-1.62.zip"
+    ])
+    validate_cliresult(result)
+    assert "is already installed" in result.output
+
+    # repository
     result = clirunner.invoke(cmd_lib, [
         "-g", "install",
         "https://github.com/platformio/platformio-libmirror.git"
@@ -136,23 +140,42 @@ def test_global_install_repository(clirunner, validate_cliresult,
 def test_global_lib_list(clirunner, validate_cliresult):
     result = clirunner.invoke(cmd_lib, ["-g", "list"])
     validate_cliresult(result)
-    assert all([n in result.output for n in ("OneWire", "DHT22", "64")])
+    assert all([
+        n in result.output for n in
+        ("Source: https://github.com/adafruit/Adafruit-ST7735-Library/archive/master.zip",
+         "Version: 5.10.1",
+         "Source: git+https://github.com/gioblu/PJON.git#3.0",
+         "Version: 1fb26fd", "RadioHead-1.62")
+    ])
 
     result = clirunner.invoke(cmd_lib, ["-g", "list", "--json-output"])
     assert all([
-        n in result.output
-        for n in (
-            "PJON", "git+https://github.com/knolleary/pubsubclient",
-            "https://github.com/bblanchon/ArduinoJson/archive/v5.8.2.zip")
+        n in result.output for n in
+        ("__pkg_dir",
+         '"__src_url": "git+https://gitlab.com/ivankravets/rs485-nodeproto.git"',
+         '"version": "5.10.1"')
     ])
     items1 = [i['name'] for i in json.loads(result.output)]
     items2 = [
-        "OneWire", "DHT22", "PJON", "ESPAsyncTCP", "ArduinoJson",
-        "PubSubClient", "rs485-nodeproto", "Adafruit ST7735 Library",
-        "RadioHead-1.62", "DallasTemperature", "NeoPixelBus",
-        "EspSoftwareSerial", "platformio-libmirror"
+        "Adafruit ST7735 Library", "ArduinoJson", "ArduinoJson", "ArduinoJson",
+        "ArduinoJson", "AsyncMqttClient", "AsyncTCP", "DallasTemperature",
+        "ESPAsyncTCP", "NeoPixelBus", "OneWire", "PJON", "PJON",
+        "PubSubClient", "RFcontrol", "RadioHead-1.62", "platformio-libmirror",
+        "rs485-nodeproto"
     ]
-    assert set(items1) == set(items2)
+    assert sorted(items1) == sorted(items2)
+
+    versions1 = [
+        "{name}@{version}".format(**item)
+        for item in json.loads(result.output)
+    ]
+    versions2 = [
+        'ArduinoJson@5.8.2', 'ArduinoJson@5.10.1', 'AsyncMqttClient@0.8.2',
+        'AsyncTCP@1.0.1', 'ESPAsyncTCP@1.1.3', 'NeoPixelBus@2.2.4',
+        'PJON@07fe9aa', 'PJON@1fb26fd', 'PubSubClient@bef5814',
+        'RFcontrol@77d4eb3f8a', 'RadioHead-1.62@0.0.0'
+    ]
+    assert set(versions1) >= set(versions2)
 
 
 def test_global_lib_update_check(clirunner, validate_cliresult):
@@ -160,7 +183,7 @@ def test_global_lib_update_check(clirunner, validate_cliresult):
         cmd_lib, ["-g", "update", "--only-check", "--json-output"])
     validate_cliresult(result)
     output = json.loads(result.output)
-    assert set(["ArduinoJson", "EspSoftwareSerial",
+    assert set(["RFcontrol",
                 "NeoPixelBus"]) == set([l['name'] for l in output])
 
 
@@ -181,11 +204,9 @@ def test_global_lib_update(clirunner, validate_cliresult):
     # update rest libraries
     result = clirunner.invoke(cmd_lib, ["-g", "update"])
     validate_cliresult(result)
-    validate_cliresult(result)
-    assert result.output.count("[Fixed]") == 5
-    assert result.output.count("[Up-to-date]") == 10
-    assert "Uninstalling ArduinoJson @ 5.7.3" in result.output
-    assert "Uninstalling EspSoftwareSerial @ 00589a3250" in result.output
+    assert result.output.count("[Fixed]") == 6
+    assert result.output.count("[Up-to-date]") == 11
+    assert "Uninstalling RFcontrol @ 77d4eb3f8a" in result.output
 
     # update unknown library
     result = clirunner.invoke(cmd_lib, ["-g", "update", "Unknown"])
@@ -207,16 +228,17 @@ def test_global_lib_uninstall(clirunner, validate_cliresult,
     # uninstall the rest libraries
     result = clirunner.invoke(cmd_lib, [
         "-g", "uninstall", "1", "https://github.com/bblanchon/ArduinoJson.git",
-        "ArduinoJson@!=5.6.7", "EspSoftwareSerial@>=3.3.1"
+        "ArduinoJson@!=5.6.7", "RFcontrol"
     ])
     validate_cliresult(result)
 
     items1 = [d.basename for d in isolated_pio_home.join("lib").listdir()]
     items2 = [
-        "ArduinoJson_ID64", "ArduinoJson_ID64@5.6.7", "DallasTemperature_ID54",
-        "DHT22_ID58", "ESPAsyncTCP_ID305", "NeoPixelBus_ID547", "PJON",
-        "PJON@src-79de467ebe19de18287becff0a1fb42d", "PubSubClient",
-        "RadioHead-1.62", "rs485-nodeproto", "platformio-libmirror"
+        "RadioHead-1.62", "rs485-nodeproto", "platformio-libmirror",
+        "PubSubClient", "ArduinoJson@src-69ebddd821f771debe7ee734d3c7fa81",
+        "ESPAsyncTCP_ID305", "DallasTemperature_ID54", "NeoPixelBus_ID547",
+        "PJON", "AsyncMqttClient_ID346", "ArduinoJson_ID64",
+        "PJON@src-79de467ebe19de18287becff0a1fb42d", "AsyncTCP_ID1826"
     ]
     assert set(items1) == set(items2)
 
@@ -248,7 +270,7 @@ def test_lib_stats(clirunner, validate_cliresult):
     validate_cliresult(result)
     assert all([
         s in result.output
-        for s in ("UPDATED", "ago", "http://platformio.org/lib/show")
+        for s in ("UPDATED", "POPULAR", "http://platformio.org/lib/show")
     ])
 
     result = clirunner.invoke(cmd_lib, ["stats", "--json-output"])