+++ tests/commands/test_platform.py
@@ -24,27 +24,25 @@ def test_search_json_output(clirunner, validate_cliresult, isolated_pio_home):
     validate_cliresult(result)
     search_result = json.loads(result.output)
     assert isinstance(search_result, list)
-    assert len(search_result)
+    assert search_result
     platforms = [item['name'] for item in search_result]
     assert "atmelsam" in platforms
 
 
-def test_search_raw_output(clirunner, validate_cliresult, isolated_pio_home):
+def test_search_raw_output(clirunner, validate_cliresult):
     result = clirunner.invoke(cli_platform.platform_search, ["arduino"])
     validate_cliresult(result)
     assert "teensy" in result.output
 
 
-def test_install_unknown_version(clirunner, validate_cliresult,
-                                 isolated_pio_home):
+def test_install_unknown_version(clirunner):
     result = clirunner.invoke(cli_platform.platform_install,
                               ["atmelavr@99.99.99"])
     assert result.exit_code == -1
     assert isinstance(result.exception, exception.UndefinedPackageVersion)
 
 
-def test_install_unknown_from_registry(clirunner, validate_cliresult,
-                                       isolated_pio_home):
+def test_install_unknown_from_registry(clirunner):
     result = clirunner.invoke(cli_platform.platform_install,
                               ["unknown-platform"])
     assert result.exit_code == -1
@@ -63,7 +61,7 @@ def test_install_known_version(clirunner, validate_cliresult,
     assert len(isolated_pio_home.join("packages").listdir()) == 1
 
 
-def test_install_from_vcs(clirunner, validate_cliresult, isolated_pio_home):
+def test_install_from_vcs(clirunner, validate_cliresult):
     result = clirunner.invoke(cli_platform.platform_install, [
         "https://github.com/platformio/"
         "platform-espressif8266.git#feature/stage", "--skip-default-package"
@@ -72,17 +70,17 @@ def test_install_from_vcs(clirunner, validate_cliresult, isolated_pio_home):
     assert "espressif8266" in result.output
 
 
-def test_list_json_output(clirunner, validate_cliresult, isolated_pio_home):
+def test_list_json_output(clirunner, validate_cliresult):
     result = clirunner.invoke(cli_platform.platform_list, ["--json-output"])
     validate_cliresult(result)
     list_result = json.loads(result.output)
     assert isinstance(list_result, list)
-    assert len(list_result)
+    assert list_result
     platforms = [item['name'] for item in list_result]
     assert set(["atmelavr", "espressif8266"]) == set(platforms)
 
 
-def test_list_raw_output(clirunner, validate_cliresult, isolated_pio_home):
+def test_list_raw_output(clirunner, validate_cliresult):
     result = clirunner.invoke(cli_platform.platform_list)
     validate_cliresult(result)
     assert all(
@@ -111,4 +109,4 @@ def test_uninstall(clirunner, validate_cliresult, isolated_pio_home):
     result = clirunner.invoke(cli_platform.platform_uninstall,
                               ["atmelavr", "espressif8266"])
     validate_cliresult(result)
-    assert len(isolated_pio_home.join("platforms").listdir()) == 0
+    assert not isolated_pio_home.join("platforms").listdir()