+++ tests/commands/test_settings.py
@@ -19,6 +19,6 @@ from platformio.commands.settings import cli
 def test_settings_check(clirunner, validate_cliresult):
     result = clirunner.invoke(cli, ["get"])
     validate_cliresult(result)
-    assert len(result.output)
+    assert result.output
     for item in app.DEFAULT_SETTINGS.items():
         assert item[0] in result.output