+++ tests/commands/test_update.py
@@ -16,11 +16,7 @@ from platformio.commands.update import cli as cmd_update
 
 
 def test_update(clirunner, validate_cliresult):
-    matches = (
-        "Platform Manager",
-        "Up-to-date",
-        "Library Manager"
-    )
+    matches = ("Platform Manager", "Up-to-date", "Library Manager")
     result = clirunner.invoke(cmd_update, ["--only-check"])
     validate_cliresult(result)
     assert all([m in result.output for m in matches])