+++ tests/conftest.py
@@ -17,10 +17,7 @@ import os
 import pytest
 from click.testing import CliRunner
 
-
-@pytest.fixture(scope="module")
-def clirunner():
-    return CliRunner()
+from platformio import util
 
 
 @pytest.fixture(scope="session")
@@ -34,6 +31,11 @@ def validate_cliresult():
 
 
 @pytest.fixture(scope="module")
+def clirunner():
+    return CliRunner()
+
+
+@pytest.fixture(scope="module")
 def isolated_pio_home(request, tmpdir_factory):
     home_dir = tmpdir_factory.mktemp(".platformio")
     os.environ['PLATFORMIO_HOME_DIR'] = str(home_dir)
@@ -43,3 +45,8 @@ def isolated_pio_home(request, tmpdir_factory):
 
     request.addfinalizer(fin)
     return home_dir
+
+
+@pytest.fixture(scope="function")
+def without_internet(monkeypatch):
+    monkeypatch.setattr(util, "_internet_on", lambda: False)